requirements.txt
pypdf
pytest
boto3
python-dotenv==1.0.1 # For reading environment variables stored in .env file
langchain==0.2.2
langchain-community==0.2.3
langchain-openai==0.1.8 # For embeddings
unstructured==0.14.4 # Document loading
openai==1.31.1 # For embeddings
tiktoken==0.7.0  # For embeddings 

pymilvus==2.5.0 #python sdk for milvus
milvus-model
langchain_upstage
wikipedia-api
torch